Remove occurrences of old scrapping code
Remove the repeated code in certain SeriesData properties. Leave the
properties which require confirmation such as 'IsCompleted'.

Add a formatting method to the Results helper class to uniformly print the
series information for unit tests.

Pre-parse the InnerHtml property in the ScrapeInformation(string, string,
string) method and replace the '<br>' element with a new line. Consider
the effects of its use and whether to allow the user to format the value
themselves.

The element provides a 'break' which in turn creates a space. With it
removed (when calling 'InnerText'), the space is no longer there. (i.e.
From Solaris-SVU:Though)