Implement the 'IsLicensed' ISeriesData property
Implement the IsLicensed property which used to determine whether
the series has been licensed in English in the SeriesData class.

Implement 'ScrapeAndCompareInformation' to compare the scraped value with
an expected value for boolean properties.

Create a public enum to store the different types of series. It will
allow for better Linq queries when checking if the series is a novel or
manga.